==> R/01_startup.R <==
#### 01 STARTUP ################################################################

# Load packages -----------------------------------------------------------

suppressPackageStartupMessages({
  library(tidyverse)
  library(lubridate)
  library(upgo)
  library(sf)
  library(stringr)
  library(tidyr)
  library(future)
  library(qs)
  library(rvest)
})


# Set global variables ----------------------------------------------------

col_palette <- 
  c("#FF6600", "#CC6699", "#3399CC", "#FFCC66", "#074387", "#6EEB83", "#008A43", 
    "#FFD500", "#A80858")
#scales::show_col(col_palette)
